backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
httpx
pydantic
torch --index-url https://download.pytorch.org/whl/cu121
transformers
accelerate
bitsandbytes
scipy
faster-whisper
optimum
